Refactor types file
