+++ react-delegationdashboard/src/pages/Dashboard/PendingUndelegated/UndelegatedValueType.ts
@@ -1,0 +1,8 @@
+export class UndelegatedValueType {
+  value: string;
+  timeLeft: number;
+  public constructor(value: string, timeLeft: number) {
+    this.value = value;
+    this.timeLeft = timeLeft;
+  }
+}
