+++ react-delegationdashboard/src/pages/Owner/Nodes/NodeRow.tsx
@@ -8,8 +8,8 @@ import { Dropdown } from 'react-bootstrap';
 import { useContext } from 'context';
 import { nodeActions } from './helpers/nodeTypes';
 import { nodeTransactions } from './helpers/stakeHooks';
-import { NodeType } from 'helpers/types';
 import { stakingContract } from 'config';
+import { NodeType } from './helpers/nodeType';
 
 type ActionType = 'unStake' | 'unJail' | 'unBond' | 'reStake' | 'stake' | 'remove';
 
@@ -21,7 +21,7 @@ const allowedActions: { [key: string]: ActionType[] } = {
   notStaked: ['stake', 'remove'],
 };
 
-const NodeRow = ({ blsKey: key, index }: { blsKey: NodeType; index: number }) => {
+const NodeRow = ({ blsKey: key }: { blsKey: NodeType; index: number }) => {
   const { explorerAddress, dapp, delegationContract } = useContext();
   const ref = React.useRef(null);
